chore: move parseBoolean to parse-utils
