+++ packages/smithy-client/src/index.ts
@@ -1,4 +1,3 @@
-export * from "./boolean";
 export * from "./client";
 export * from "./command";
 export * from "./document-type";
@@ -7,6 +6,7 @@ export * from "./extended-encode-uri-component";
 export * from "./get-array-if-single-item";
 export * from "./get-value-from-text-node";
 export * from "./lazy-json";
+export * from "./parse-utils";
 export * from "./date-utils";
 export * from "./split-every";
 export * from "./constants";
