+++ packages/smithy-client/src/parse-utils.spec.ts
@@ -1,4 +1,4 @@
-import { parseBoolean } from "./boolean";
+import { parseBoolean } from "./parse-utils";
 
 describe("parseBoolean", () => {
   it('Returns true for "true"', () => {
